bugfixing

+++ models/expense.go
@@ -13,7 +13,7 @@ type Expense struct {
 	ProjectID  primitive.ObjectID `json:"projectId,omitempty" bson:"projectId,omitempty"`
 	CategoryID primitive.ObjectID `json:"categoryId,omitempty" bson:"_categoryId,omitempty"`
 	Title      string             `json:"title,omitempty" bson:"title,omitempty"`
-	Amount     float32            `json:"amount,omitempty" bson:"amount,omitempty"`
+	Amount     float64            `json:"amount,omitempty" bson:"amount,omitempty"`
 	AddedDate  time.Time          `json:"addedDate,omitempty" bson:"addedDate,omitempty"`
 	Category   Category           `json:"category,omitempty" bson:"-"`
 }
